+++ beacon/validator/src/main/java/tech/pegasys/teku/validator/coordinator/Eth1DataProvider.java
@@ -18,8 +18,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import org.apache.commons.lang3.tuple.Pair;
+import tech.pegasys.teku.ethereum.pow.api.DepositTreeSnapshot;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
 import tech.pegasys.teku.spec.datastructures.metadata.StateAndMetaData;
@@ -62,6 +64,10 @@ public class Eth1DataProvider {
     final UInt64 slotsLeft = startSlot.plus(votingSlots).minus(beaconState.getSlot());
     final UInt64 votesRequired = votingSlots.dividedBy(2);
     return new VotingPeriodInfo(votesRequired, votingSlots, slotsLeft);
+  }
+
+  public Optional<DepositTreeSnapshot> getFinalizedDepositTreeSnapshot() {
+    return depositProvider.getFinalizedDepositTreeSnapshot();
   }
 
   public static class VotingPeriodInfo {