+++ CHANGELOG.md
@@ -20,6 +20,7 @@ For information on changes in released versions of Teku, see the [releases page]
  - Optimised storage of latest vote information by batching updates
  - Ensured dependencies are up to date
  - Validator Registration signature integration with external signer
+ - Teku-specific Beacon Node API on `/teku/v1/beacon/deposit_snapshot` providing finalized Deposit Tree Snapshot according to the draft EIP-4881
 
 ### Bug Fixes
  - Fix not rendering emoticons correctly in graffiti when running in a Docker container